chore: adjust node versions to support multiple Nextcloud versions
This is a library and should support multiple Nextcloud versions.
E.g. Nextcloud 30 apps are bound to Node 20, current apps Node 22 and
apps in a couple of month Node 24.
Thus - and because this is not a node but browser library - we should
support all LTS Node versions in this range.

Instead we should just require specific development versions.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>